backend/pyproject.toml
# backend/pyproject.toml
[project]
name = "securepass-backend"
version = "0.1.0"
description = "SecurePass - Zero-Knowledge File Sharing Platform Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # FastAPI関連
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    # Prisma ORM
    "prisma>=0.11.0",
    # 認証・セキュリティ
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=41.0.7",
    "PyJWT>=2.8.0",
    # AWS S3 (Cloudflare R2)
    "boto3>=1.34.14",
    "aioboto3>=12.1.0",
    # Redis
    "redis>=5.0.1",
    "celery>=5.3.4",
    # 環境変数
    "python-dotenv>=1.0.0",
    # ユーティリティ
    "httpx>=0.25.2",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dateutil>=2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.4",
    "pytest-asyncio>=0.21.1",
    "black>=23.12.1",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
    "pytest-cov>=4.1.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
asyncio_mode = "auto"
